// lib/supabase/database.types.ts
export interface Database {
  public: {
    Tables: {
// ...
      product_variants: {
        Row: {
          id: string
          product_id: string
          size_id: string
          inventory_count: number
          sku: string | null
          price: number | null
          created_at: string
          updated_at: string
        }
        Insert: {
          id?: string
          product_id: string
          size_id: string
          inventory_count?: number
          sku?: string | null
          price?: number | null
          created_at?: string
          updated_at?: string
        }
        Update: {
          id?: string
          product_id?: string
          size_id?: string
          inventory_count?: number
          sku?: string | null
          price?: number | null
          created_at?: string
          updated_at?: string
        }
      }
// ...
      transactions: {
        Row: {
          id: string
          product_id: string
          variant_id: string
          price: number
          quantity: number
          buyer_id: string
          department: string
          province: string
          address: string
          status: string
          payment_method: string
          payment_status: string
          created_at: string
          updated_at: string
        }
        Insert: {
          id?: string
          product_id: string
          variant_id: string
          price: number
          quantity?: number
          buyer_id: string
          department: string
          province: string
          address: string
          status?: string
          payment_method: string
          payment_status?: string
          created_at?: string
          updated_at?: string
        }
        Update: {
          id?: string
          product_id?: string
          variant_id?: string
          price?: number
          quantity?: number
          buyer_id?: string
          department?: string
          province?: string
          address?: string
          status?: string
          payment_method?: string
          payment_status?: string
          created_at?: string
          updated_at?: string
        }
      }
    }
// ...
  }
}